+++ Etomo/src/etomo/comscript/ExtractmagradParam.java
@@ -68,7 +68,7 @@ public class ExtractmagradParam {
     this.gradientTable = gradientTable;
   }
 
-  public final void setRotationAngle(float rotationAngle) {
+  public final void setRotationAngle(double rotationAngle) {
     this.rotationAngle.set(rotationAngle);
   }
 
@@ -85,6 +85,9 @@ public class ExtractmagradParam {
 }
 /**
  * <p> $Log$
+ * <p> Revision 1.7  2011/02/21 21:26:24  sueh
+ * <p> bug# 1437 Reformatting.
+ * <p>
  * <p> Revision 1.6  2010/02/17 04:47:54  sueh
  * <p> bug# 1301 Using the manager instead of the manager key do pop up
  * <p> messages.
